backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
librosa==0.10.1
numpy==1.26.4
scipy==1.12.0
soundfile==0.12.1
mutagen==1.47.0
stripe==8.5.0
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.1
# Metadata integrations
acoustid==1.2.2
musicbrainzngs==0.7.1
spotipy==2.23.0
pylast==5.3.0

